trabajando en tabla balance

+++ src/components/Dashboard/DailyExpenses/TableSection/TableExpenses.jsx
@@ -4,42 +4,65 @@ import { transformColumns } from "@/helpers/helpers";
 import { CloseOutlined } from "@ant-design/icons";
 import ModalCharge from "../../Transactions/ModalCharge";
 
-const TableExpenses = ({ dataSource, setDataSource, refreshData, filterType }) => {
+const TableExpenses = ({
+  dataSource,
+  setDataSource,
+  refreshData,
+  filterType,
+}) => {
   const [isModalOpen, setIsModalOpen] = useState(false);
   const [selectedExpense, setSelectedExpense] = useState(null);
 
   const onEdit = (record) => {
     setIsModalOpen(true);
-    setSelectedExpense(record); 
+    setSelectedExpense(record);
   };
 
   const onDelete = async (record) => {
-    try {
-      const res = await fetch(`/api/expenses/${record.expense_id}`, {
-        method: "DELETE",
-        headers: {
-          "Content-Type": "application/json",
-        },
-      });
-      if (res.ok) {
-        setDataSource((prevData) =>
-          prevData.filter((item) => item.expense_id !== record.expense_id)
-        );
-        console.log("Eliminando...", record);
+    if (filterType === "expenses") {
+      try {
+        const res = await fetch(`/api/expenses/${record.expense_id}`, {
+          method: "DELETE",
+          headers: {
+            "Content-Type": "application/json",
+          },
+        });
+        if (res.ok) {
+          setDataSource((prevData) =>
+            prevData.filter((item) => item.expense_id !== record.expense_id)
+          );
+          console.log("Eliminando...", record);
+        }
+      } catch (error) {
+        console.log(error);
       }
-    } catch (error) {
-      console.log(error);
     }
-  };  
+    if (filterType === "incomes") {
+      try {
+        const res = await fetch(`/api/incomes/${record.income_id}`, {
+          method: "DELETE",
+          headers: {
+            "Content-Type": "application/json",
+          },
+        });
+        if (res.ok) {
+          setDataSource((prevData) =>
+            prevData.filter((item) => item.income_id !== record.income_id)
+          );
+          console.log("Eliminando...", record);
+        }
+      } catch (error) {
+        console.log(error);
+      }
+    }
+  };
 
   const columns = transformColumns(dataSource, onEdit, onDelete);
 
   const handleCancel = () => {
     setIsModalOpen(false);
-    setSelectedExpense([]); 
+    setSelectedExpense([]);
   };
-
-  
 
   return (
     <>
@@ -53,14 +76,16 @@ const TableExpenses = ({ dataSource, setDataSource, refreshData, filterType }) =
                 margin: 0,
               }}
             >
-               {record[`${filterType.slice(0, -1)}_description`] + " - " + record[`${filterType.slice(0, -1)}_notes`]}
+              {record[`${filterType.slice(0, -1)}_description`] +
+                " - " +
+                record[`${filterType.slice(0, -1)}_notes`]}
             </p>
           ),
         }}
         rowKey={`${filterType.slice(0, -1)}_id`}
         scroll={{ x: "100vh", y: "calc(100vh - 20rem)" }}
         bordered
-        pagination={false} 
+        pagination={false}
       />
       <Modal
         closeIcon={<CloseOutlined className="text-white" />}
@@ -68,7 +93,12 @@ const TableExpenses = ({ dataSource, setDataSource, refreshData, filterType }) =
         onCancel={handleCancel}
         footer={null}
       >
-        <ModalCharge onClose={handleCancel} refreshData={refreshData} expense={selectedExpense} filterType={filterType}/>
+        <ModalCharge
+          onClose={handleCancel}
+          refreshData={refreshData}
+          expense={selectedExpense}
+          filterType={filterType}
+        />
       </Modal>
     </>
   );
